% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sequence.R
\name{create_analyst_seq}
\alias{create_analyst_seq}
\title{Create a sequence table for AB Sciex Analyst}
\usage{
create_analyst_seq(object)
}
\arguments{
\item{object}{Sequence object containing all the information.}
}
\value{
Returns a data frame
}
\description{
Create a sequence table for AB Sciex Analyst.
}
\author{
Rico Derks
}
