% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sequence.R
\name{create_hystar_seq}
\alias{create_hystar_seq}
\title{Create a sequence table for Bruker HyStar}
\usage{
create_hystar_seq(object)
}
\arguments{
\item{object}{Sequence object containing all the information.}
}
\value{
Returns a data frame
}
\description{
Create a sequence table for Bruker Hystar.
}
\author{
Rico Derks
}
