% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sequence.R
\name{create_sample_vials}
\alias{create_sample_vials}
\title{Create a sequence with sample vial positions}
\usage{
create_sample_vials(
  start_vial = NULL,
  number_samples = NULL,
  vial_type = c("48", "96")
)
}
\arguments{
\item{start_vial}{The starting position of the vial.}

\item{number_samples}{The number of samples.}

\item{vial_type}{Define the number of vials per plate. 48 is glass vials and
96 is 96-well plate.}
}
\value{
Returns character vector with all possible vial positions.
}
\description{
Create a sequence with sample vial positions. This is needed for the
    LC which is combined with the Bruker MS. It in general has 3 colored positions
    (red, green and blue) and in there 48 vial positions.
}
\author{
Rico Derks
}
