% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sequence.R
\name{prepare_sequence_table}
\alias{prepare_sequence_table}
\title{Prepare sequence table}
\usage{
prepare_sequence_table(info)
}
\arguments{
\item{info}{Sequence object containing all the information.}
}
\value{
Returns a data frame
}
\description{
Prepare the sequence table
}
\details{
sequence
}
\author{
Rico Derks
}
